requirements.txt
pillow
numpy
requests
spotipy
RGBMatrixEmulator
nyct-gtfs
